refactor core internals to more granual file structure
