+++ packages/core/src/modules.ts
@@ -5,7 +5,7 @@ import * as PluginManager from './plugins';
 import * as ProjectManager from './projects';
 import * as ConfigManager from './configs';
 import * as SchemaManager from './schema/zod';
-import * as RuntimeManager from './runtime';
+import * as RuntimeManager from './context/runtime';
 import * as TemplateManager from './templates';
 import * as TaskManager from './tasks';
 // SUB-MODULES
