+++ packages/engine-core/src/tasks/task.rnv.project.configure.ts
@@ -33,6 +33,7 @@ import {
     findSuitableTask,
     RnvTaskFn,
     RnvContext,
+    generatePlatformAssetsRuntimeConfig,
 } from '@rnv/core';
 
 const checkIsRenativeProject = (c: RnvContext) =>
@@ -64,7 +65,7 @@ export const taskRnvProjectConfigure: RnvTaskFn = async (c, parentTask, originTa
     if (c.program.only && !!parentTask) {
         await configureRuntimeDefaults(c);
         await executeTask(c, TASK_APP_CONFIGURE, TASK_PROJECT_CONFIGURE, originTask);
-        await generateRuntimeConfig(c);
+        await generatePlatformAssetsRuntimeConfig(c);
         return true;
     }
 
@@ -120,7 +121,7 @@ export const taskRnvProjectConfigure: RnvTaskFn = async (c, parentTask, originTa
         if (!c.buildConfig.platforms) {
             await updateRenativeConfigs(c);
         }
-        await generateRuntimeConfig(c);
+        await generatePlatformAssetsRuntimeConfig(c);
         await overrideTemplatePlugins(c);
         // NOTE: this is needed to ensure missing rnv plugin sub-deps are caught
         await checkForPluginDependencies(c);
